feat: integrate react-router for navigation and separate game and game creation pages
