+++ package.json
@@ -17,6 +17,7 @@
     "rand-seed": "^2.1.7",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
+    "react-router-dom": "^7.3.0",
     "socket.io": "^4.8.1",
     "socket.io-client": "^4.8.1"
   },
